+++ src/main/java/com/example/jjangushrine/domain/cart/controller/CartController.java
@@ -1,6 +1,6 @@
-package com.example.jjangushrine.domain.order.controller;
+package com.example.jjangushrine.domain.cart.controller;
 
-import com.example.jjangushrine.domain.order.service.CartService;
+import com.example.jjangushrine.domain.cart.service.CartService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
