+++ src/main/java/com/example/jjangushrine/domain/cart/service/CartService.java
@@ -1,0 +1,7 @@
+package com.example.jjangushrine.domain.cart.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CartService {
+}
