--- src/main/java/com/example/jjangushrine/domain/order/entity/Cart.java
@@ -1,24 +1,0 @@
-package com.example.jjangushrine.domain.order.entity;
-
-import com.example.jjangushrine.domain.user.entity.User;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "carts")
-public class Cart {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long cartId;
-
-	@OneToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
-
-
-}
