--- src/main/java/com/example/jjangushrine/domain/order/service/CartServiceImpl.java
@@ -1,4 +1,0 @@
-package com.example.jjangushrine.domain.order.service;
-
-public class CartServiceImpl {
-}
